Separate `PosixError` from `SysError`
Most of this is a big `throw SysError` -> `throw PosixError` sed. This
is a rather pure-churn change I would like to get out of the way. **The
intersting part is `src/libutil/error.hh`.**

On Unix, we will only throw the `PosixError` concrete class, which has
the same constructors that `SysError` used to have.

On Windows, we will throw `WinError` *and* `PosixError`. `WinError`
(which will be created in a later PR), will use a `DWORD` instead of
`int` error value, and `GetLastError()`, which is the Windows equivalent
of the `errno` machinery. Windows will *also* use `PosixError` because
Window's "libc" (MSVCRT) implements the POSIX interface, and we use it
too.

As the docs describe, while we *throw* one of the 3 choices above (2
concrete classes or the alias), we should always *catch* `SysError`.
This ensures no matter how the implementation changes for Windows (e.g.
between `PosixError` and `WinError`) the catching logic stays the same
and stays correct.

Co-Authored-By volth <volth@volth.com>